requirements.txt
aiofiles==23.2.1
aiogram==3.2.0
aiohttp==3.9.1
aiosignal==1.3.1
annotated-types==0.6.0
async-timeout==4.0.3
attrs==23.1.0
certifi==2023.7.22
charset-normalizer==3.3.1
frozenlist==1.4.0
fuzzywuzzy==0.18.0
idna==3.4
Levenshtein==0.23.0
magic-filter==1.0.12
multidict==6.0.4
pydantic==2.5.3
pydantic_core==2.14.6
python-dotenv==1.0.0
python-Levenshtein==0.23.0
rapidfuzz==3.5.2
requests==2.31.0
typing_extensions==4.10.0
urllib3==2.2.1
yarl==1.9.2
